project-openweather/streamlit/requirements.txt
pymongo
streamlit
pandas
matplotlib
plotly
python-dotenv
